+++ Changelog.rst
@@ -5,6 +5,11 @@ Version NEXTVERSION
 
 * Write UGRID datasets with `cfdm.write`
   (https://github.com/NCAS-CMS/cfdm/issues/271)
+* Fix bug in `cfdm.write` when writing identical coordinates that have
+  different ``formula_terms``
+  (https://github.com/NCAS-CMS/cfdm/issues/380). New unit test added.
+
+----
 
 Version 1.13.0.0
 ----------------